third_party/WebKit/LayoutTests/media/mediasession/mediaimage-expected.txt

CONSOLE WARNING: line 6: dummy log entry to avoid an empty -expected.txt in virtual/
This is a testharness.js-based test.
FAIL Test that MediaImage is constructed using a dictionary MediaImage is not defined
FAIL Test the different values allowed in MediaImage init dictionary MediaImage is not defined
FAIL Test the default values for MediaImage MediaImage is not defined
FAIL Test that passing unknown values to the dictionary is a no-op MediaMetadata is not defined
Harness: the test ran to completion.
